requirements.txt
python-pptx
slugify
scipy
natsort
